requirements.txt
# Core dependencies
langchain==0.2.0
langchain-anthropic==0.1.0
langchain-community==0.0.20
langchain-text-splitters==0.3.0
anthropic==0.18.1

# Data processing
pandas>=2.2.0
numpy>=1.24.0

# Vector store and embeddings
chromadb==0.4.22

# Document processing
pypdf==3.17.4
python-docx==1.1.0
markdown==3.5.2

# Configuration and utilities
pyyaml==6.0.1
python-dotenv==1.0.0
click==8.1.7

# Web application
flask==3.0.0

# Development and testing
pytest==7.4.3
black==23.11.0
flake8==6.1.0 